+++ src/ppu.h
@@ -7,7 +7,7 @@
 #define NES_SCR_HEIGHT (240)
 
 extern void ppu_reset(void);
-extern void ppu_step(unsigned pputicks);
+extern void ppu_step(short pputicks);
 extern void ppu_log_state(void);
 extern void ppu_write(uint8_t val, uint16_t addr);
 extern uint8_t ppu_read(uint16_t addr);
